requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
boto3==1.24.19
botocore==1.27.19
Django==4.0.5
django-cleanup==6.0.0
django-cors-headers==3.13.0
django-filter==22.1
django-phone-field==1.8.1
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
jmespath==1.0.1
Pillow==9.1.1
PyJWT==2.4.0
python-dateutil==2.8.2
pytz==2022.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
